Update pid type and fix windows err

+++ robotgo.go
@@ -987,5 +987,5 @@ func FindIds(name string) ([]int32, error) {
 
 // ActivePID window active by PID
 func ActivePID(pid int32) {
-	C.active_PID(C.int32(pid))
+	C.active_PID(C.uintptr(pid))
 }
